+++ src/main/java/Command/MarkCommand.java
@@ -1,18 +1,25 @@
-public class UnMarkCommand {
+package Command;
+
+import Duke.Ui;
+import Task.Task;
+import Task.TaskList;
+
+public class MarkCommand {
+
     private int number;
-    public UnMarkCommand(String number) {
+    public MarkCommand(String number) {
         this.number = Integer.parseInt(number) - 1;
     }
 
-    public void unmark() {
+    public void mark() {
         try {
             Task curr = TaskList.get(number);
-            curr.unmark();
-            Ui.printUnmark(curr);
+            curr.mark();
+            Ui.printMark(curr);
         } catch (Exception m){
             Ui.printWrongNumber();
             String s = Ui.readCommand();
-            new UnMarkCommand(s).unmark();
+            new MarkCommand(s).mark();
         }
     }
 }
