+++ src/main/java/Duke/Parser.java
@@ -1,5 +1,5 @@
+package Duke;
 
-import java.util.Scanner;
 public class Parser {
 
     public String[] parse(String command) {
