+++ src/main/java/Task/Task.java
@@ -1,7 +1,8 @@
+package Task;
+
 import java.io.File;
 import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
-import java.util.Scanner;
+
 
 public abstract class Task {
